use of functions
